+++ src/transforms/lua/v1/mod.rs
@@ -218,6 +218,13 @@ impl rlua::UserData for LuaEvent {
     }
 }
 
+pub fn format_error(error: &rlua::Error) -> String {
+    match error {
+        rlua::Error::CallbackError { traceback, cause } => format_error(&cause) + "\n" + traceback,
+        err => err.to_string(),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::{format_error, Lua};
